Make dependencies that are only required to build the tool optional.
The rational is that users of the library should not have to download and compile unncessary dependencies.

Another issue is that the version of `meval` uses an extremely old version of `nom` and the PR for the `meval` project which updates `meval` to use `nom` 7.x is stalled for almost a year now.

All of this is prompted by the fact that compiler warnings are generated by the old `meval` and `nom` code which can be avoided if the dependencies of the library are optional.